[IDLE-275] 요양보호사(다른사람) 프로필 ViewModel 구현
UseCase 주입작업을 진행할 예정입니다.
